+++ handlers/discount-api/test/eligibilityChecker.test.ts
@@ -4,13 +4,11 @@ import {
 	itemsForSubscription,
 	toSimpleInvoiceItems,
 } from '@modules/zuora/billingPreview';
-import {
-	billingPreviewSchema,
-	zuoraSubscriptionResponseSchema,
-} from '@modules/zuora/zuoraSchemas';
+import { zuoraSubscriptionResponseSchema } from '@modules/zuora/zuoraSchemas';
 import { ZuoraCatalogHelper } from '@modules/zuora-catalog/zuoraCatalog';
 import { zuoraCatalogSchema } from '@modules/zuora-catalog/zuoraCatalogSchema';
 import dayjs from 'dayjs';
+import { billingPreviewSchema } from '../../../modules/zuora/src/types/objects/billingPreview';
 import catalogJsonProd from '../../../modules/zuora-catalog/test/fixtures/catalog-prod.json';
 import {
 	EligibilityChecker,
